refactor: 카테고리 entity 위치 변경
