+++ src/main/java/shop/nuribooks/books/repository/category/CategoryRepository.java
@@ -2,7 +2,7 @@ package shop.nuribooks.books.repository.category;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import shop.nuribooks.books.entity.book.category.Category;
+import shop.nuribooks.books.entity.book.Category;
 
 public interface CategoryRepository extends JpaRepository<Category, Long> {
 	boolean existsByNameAndParentCategoryIsNull(String name);
