+++ src/test/java/shop/nuribooks/books/controller/category/CategoryControllerTest.java
@@ -19,7 +19,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import shop.nuribooks.books.controller.GlobalExceptionHandler;
 import shop.nuribooks.books.dto.category.CategoryRequest;
 import shop.nuribooks.books.dto.category.CategoryResponse;
-import shop.nuribooks.books.entity.book.category.Category;
+import shop.nuribooks.books.entity.book.Category;
 import shop.nuribooks.books.exception.category.CategoryAlreadyExistException;
 import shop.nuribooks.books.exception.category.CategoryNotFoundException;
 import shop.nuribooks.books.service.category.CategoryService;
